Move into tools directory
