--- docs/detection/smoother.md
@@ -1,3 +1,0 @@
-## Smoother
-
-:::supervision.detection.smoother.Smoother
