+++ docs/detection/tools/smoother.md
@@ -1,0 +1,3 @@
+## Detection Smoother
+
+:::supervision.detection.tools.smoother.Smoother
