+++ mkdocs.yml
@@ -37,12 +37,12 @@ nav:
         - Core: classification/core.md
     - Detections:
         - Core: detection/core.md
-        - Smoother: detection/smoother.md
         - Utils: detection/utils.md
     - Tools:
         - Line Zone: detection/tools/line_zone.md
         - Polygon Zone: detection/tools/polygon_zone.md
         - Inference Slicer: detection/tools/inference_slicer.md
+        - Detection Smoother: detection/tools/smoother.md
     - Annotators: annotators.md
     - Trackers: trackers.md
     - Datasets: datasets.md
